Added walk_mut in preparation for `store` implementations
The disadvantage of this generics approach, or a macro approach like the
crate `duplicate`, is that the output code is duplicated. And these are
hot-path functions `walk`/`walk_mut`. And stores and loads are likely to
occur very close to each other in RISC-V user apps. Doesn't sound like
good news for the instruction cache.