Switched to partially using snow
It is not possible to use snow/Noise with an in-memory incrementing
nonce, as this assumes an underlying reliable transport layer. The
alternative is to use the packet number in the QUIC header as a nonce
(which can arrive out of order).

Unfortunately, this means essentially reimplementing the
nonce/encryption parts of Noise during the transport phase. Other
projects that have integrated Noise with QUIC have done this, though.

Also, this commit enables the risky-raw-split feature of snow, in
preparation for passing the key material into the reimplementation
(which I agree is risky and unfortunate), and is not used by this
commit yet, only future commits.