Fixed incorrect expectation of Noise message length by sending length
Unfortunately, although QUIC/Quinn sets the length and offset for us of
the data in a CRYPTO frame (the part of the Noise message that fits in a
UDP datagram), there isn't a field for the total length of a single
Noise message, as that is sort of up to us.

If our Noise messages were small, we'd be able to use the existing
information and assume it fits into one datagram, however post-quantum
keys and KEM ciphertexts are huge and require multiple datagrams.

Also, for some reason this commit cuts the instruction count of
read_handshake by 45%.